More code cleanup!
Functions for getting one or multiple posts from db, as well as adding and editing posts in the db. Function for "constructing" new Posts with "alternate constructor" (basically a function that makes up for js not having overloading 😭)

It was time lol, some of those functions were becoming interdependent messes.