Data Caching!!!!
To prevent extraneous firestore reads, the website reads from the db only once on load and caches all post data in sessionStorage as JSON.

It then reads this data throughout the site.

Further, the usage of URL tracking while navigating the site allows for sessionStorage to be refreshed to match firestore whenever a user refreshes on only the home page.